msfsm/backend/pyproject.toml
[tool.poetry]
name = "msfsm-api"
version = "0.1.3"
description = "API for the creation, compilation and deployment of multi-scale finite state machine-based smart contracts"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.115.12"}
typer = "^0.15.3"
python-dotenv = "^1.1.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.4.0"}
pytest = "^8.3.5"
# python-multipart is already included in fastapi[all]
pydantic-settings = "^2.9.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
